Extract parallel batch execution builtin
