+++ src/builtins/index.ts
@@ -1,2 +1,3 @@
 export * from './command-execution.builtin';
+export * from './parallel-batches';
 export * from './standard-setup';
