+++ src/builtins/parallel-batches/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @packageDocumentation
+ * @module run-z
+ */
+export * from './parallel-batch.builtin';
+export * from './parallel-batch.rule';
