+++ src/core/batches/index.ts
@@ -4,4 +4,3 @@ export * from './batch-planner';
 export * from './batch-rule';
 export * from './batcher';
 export * from './batching';
-export * from './rules';
