--- src/core/batches/rules/index.ts
@@ -1,1 +1,0 @@
-export * from './parallel-batch.rule';
